Add ContactInfo route and add react-routes file
