+++ app/assets/angular/header/header.tpl.html
@@ -26,6 +26,9 @@
         <li class="nav-item" ui-sref-active="active">
           <a class="nav-link" ui-sref="urlFeedbacks"> Feedback </a>
         </li>
+        <li class="nav-item" ui-sref-active="active">
+          <a class="nav-link" ui-sref="urlContactInfo"> Contact Info </a>
+        </li>
       </ul>
     </div>
   </nav>
